Refactor collectionIndexes to start at 0
I believe that this will be the most intuitive way of indexing
collections for developers, and fits the defaults of JavaScript.

Determining the indexes for a collection requires looking at the JSON
representation of a front page, so for the time being it seems
reasonable to assume that people editing this config will be familiar
with the zero-index convention.