Remove 'palette styles' collections from config
These collections were included because the original configs were
targetting the first collection on the page, but the first collection
just contains data on palette styles; it's not a real collection of
stories. Removing it now because we've documented that there was an
intention to use the first _actual_ collection on the page, which will
be included in the general config review that's coming up.