Add '-old' lambdas for parity with INFRA
Currently we're provisioning both the 'old' lambdas and the new ones in
parallel. Ideally we wouldn't need to create four for CODE as well, but
because they're provisioned as a list of deployments in riff-raff.yaml,
we need to make sure that there's parity between CODE and INFRA because
they use the same riff-raff.yaml file.