Replace workspace `format` cmds with root cmd
Files in the root weren't getting formatted, and some files in the
packages were being missed, too. Prettier seems to work fine for all of
our needs in this project at present, so we'll get better coverage if we
run it on everything that isn't gitignored. It doesn't seem to have a
significant performance penalty from my local testing.