Decrease alarm sensitivity
This change makes the alarm require errors in 3 evaluation periods
rather than 2. We have recently seen a couple of instances of false
alarms where the APIs pressreader depends on are unavailable for >30
minutes with no "business value" impact on this service.

We should make the alarms less sensitivity to reduce the risk of
ignoring things in the future if these periods of temporary
unavailability continue.