Remove unmaintained front link
The `us/sustainable-business` URL has been erroring, preventing the
lambda from running. However, on investigation it looks like the most
recent stories on this Front are from 2017, and on the web you get
redirected to a different page, so it looks very likely that this front
is no longer maintained.