Update lambda schedules for switch-over
Increase the frequency of the 'new' lambdas so that they run every hour.
Reduce the frequency of the 'old' lambdas to once every 4 hours.

This is in anticipation of Pressreader switching over from using the
legacy access point to using the new API.

Once we've confirmed that the switch is working as planned, we can
remove the 'old' lambdas entirely.